refactor organization of code to fix dabao target
kpc isn't possible on dabao due to the pins not being broken out,
so gate out the service entirely for that target.
