+++ services/bao1x-hal-service/src/servers.rs
@@ -1,3 +1,4 @@
+#[cfg(feature = "board-baosec")]
 pub mod keyboard;
 
 // TRNG server is only needed for baosec; dabao can cheat and use the kernel TRNG port
